% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protoc.R
\name{download_protoc}
\alias{download_protoc}
\title{Download precompiled protoc binaries
' This function downloads precompiled \code{protoc} binaries for a specified version.}
\usage{
download_protoc(
  version = defaultDownloadVersion(),
  dest_file_path = basename(bin_file_name_extract())
)
}
\arguments{
\item{version}{A character string specifying the version of \code{protoc} to download. Default is "33.0".}
}
\value{
The path to the downloaded \code{protoc} binary.
}
\description{
Download precompiled protoc binaries
' This function downloads precompiled \code{protoc} binaries for a specified version.
}
\examples{
\dontrun{
protoc_path <- download_protoc("33.0")
}
}
