% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protoc.R
\name{generate_go_from_proto}
\alias{generate_go_from_proto}
\title{Generate Go code from a proto file}
\usage{
generate_go_from_proto(
  proto_file,
  go_out_dir,
  protoc_path = NULL,
  proto_path = NULL,
  protoc_gen_go_version = NULL
)
}
\arguments{
\item{proto_file}{Path to the .proto file.}

\item{go_out_dir}{Output directory for Go code.}

\item{protoc_path}{Path to the protoc binary (optional, will use download_protoc() if not provided).}

\item{proto_path}{Directory to use as --proto_path (optional, defaults to dirname(proto_file)).}

\item{protoc_gen_go_version}{Version of protoc-gen-go to use (default: same as protoc version).}
}
\value{
Path to the generated Go file(s).
}
\description{
Generate Go code from a proto file
}
