% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protoc.R
\name{get_protoc_version}
\alias{get_protoc_version}
\title{Get protoc version
' This function retrieves the version of the installed \code{protoc} binary.}
\usage{
get_protoc_version(protoc_path)
}
\arguments{
\item{protoc_path}{The path to the \code{protoc} binary.}
}
\value{
A character string representing the version of \code{protoc}
}
\description{
Get protoc version
' This function retrieves the version of the installed \code{protoc} binary.
}
