% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protoc.R
\name{install_protoc_gen_go}
\alias{install_protoc_gen_go}
\title{Install protoc-gen-go plugin}
\usage{
install_protoc_gen_go(version = "latest", out_path = NULL)
}
\arguments{
\item{version}{The version of protoc-gen-go to install (default: latest)}

\item{out_path}{The full path where to install the protoc-gen-go executable (optional, default: $GOPATH/bin or Go's default bin dir)}
}
\value{
The path to the installed protoc-gen-go executable
}
\description{
Install protoc-gen-go plugin
}
